app/services/benchmarks/tts_stt_eval/requirements.txt
jiwer
pandas
tqdm
python-dotenv
requests
